feat: simplify function pipeline by removing dynamic configuration and move sidebar items to shared utility.
